Rename KoiosClient module name to koios_client
