+++ backend/app/koios_client/models.py
@@ -1,21 +1,22 @@
+from pydantic import BaseModel
 
 
 # Response models for history endpoints
-class ChatMessage:
+class ChatMessage(BaseModel):
     """A single chat message."""
 
     role: str
     content: str
 
 
-class ChatHistoryResponse:
+class ChatHistoryResponse(BaseModel):
     """Response model for chat history."""
 
     history: list[dict[str, str]]
     message_count: int
 
 
-class ClearHistoryResponse:
+class ClearHistoryResponse(BaseModel):
     """Response model for clearing history."""
 
     messages_deleted: int
